[BugFix] FF awae: set v_plane=0 when center of wake plane leaves domain
If a wake plane center leaves the domain, set the velocity of the plane to zero.  In low wind speed cases where the center of a skewed wake plane was leaving domain, the remaining sampled region of the plane could get very small and have negative weighting.  If this happens with a bit of flow reversal, the plane could suddenly jump backwards by a substantial distance.  This was leading to some strange looking results and odd interactions with turbines near that edge of the domain.